Prevent GTM loading until consent is set
When a user first lands on the website the default consent of 'block
everything' is applied prior to loading the GTM container. Once they opt-in/out
of cookies we update GTM with the new consent level and tags will fire on
triggers thereafter.

This is an issue for the first page view, as the page view event triggers
_before_ the consent is set by the user. Instead, we want to delay loading the
GTM container until the initial consent is set. This results in the page view
event firing _after_ the consent is applied, even though the page view has
already happened.