Remove privacy policy step from mailing list sign up
We no longer require an explicit acceptance checkbox for the privacy policy
during mailing list sign up. Instead, we shift the logic to the first step so
the latest policy is attached to the sign up.

There can now be two potential 'completion steps' in the sign up journey; the
postcode step and, if an existing candidate has their postcode pre-filled
(causing the postcode step to be skipped) then their journey will complete on
the subject step.

We need to default the policy ID to the latest to account for the magic link
sign up flow, which skips the first step (that would otherwise obtain and
specify the privacy policy ID). Going forward we will look to remove the magic
link code as this is no longer used/needed.