Remove privacy policy checkbox from event sign up
Remove the privacy policy checkbox from the event sign up. As the two
post-authentication steps are now both potentially optional this can result in
a very short sign up journey for existing candidates that have a telephone
already in the CRM and are subscribed to the mailing list. They will
essentially authenticate then see the completion screen. The button on the
authentication step will read 'next step' instead of 'complete sign up' as we
can't know until they've submitted if we need to show them additional steps or
not.