Change site name in structured data
Google is not registering out site name from the structure data. As far as we
can tell its in the correct format, so we think Google may be ignoring our site
name because it things we're trying to impersonate a government body (which is
one of the reasons they list).

Remove `GOV.UK` from the site name to see if Google registers it.