Fix colour contrast warning on WAVE tool
The WAVE chrome extension reports a colour contrast issue
on the steps heading/text; it incorrectly thinks we're
rendering white text on a white background.

It also picks up on a contrast issue with the skip to
main content link, but as this is hidden until you focus
on it it's a non-issue (after focusing there is no colour
contrast problem).

We can silence this warning by giving the elemnts an
explicit purple background.