Forward query params to cookie preferences
Users will often arrive on the website with tracking
query parameters in the URL. As no tags fire until they
accept cookies we find that if they go to cookie
preferences and then accept the original tracking query
parameters are lost.

Forward query parameters to the cookie preferences page,
so that if the user then accepts cookies we get the
tracking information firing.