Handle when there are no callback slots available
Currently if there are no callback slots available in the CRM the API will
return a default/fallback list of slots. Going forward we want to prevent
users from booking a callback if the slots have been exhausted.

Remove the 'Book a callback' button on the mailing list completion page when
there is no availability.

Prevent the user from proceeding past the callback step of the sign up
journey if they happen to be mid-way into the journey when the slots are all
used up. Instead, we will display a message encouraging them to call us.