Add QuoteListComponent
We want to display a list of quotes on some of our content pages; this is
similar to the `PhotoQuoteListComponent` but without the photos and with
the heading half overlapping the containing box. I initially tried to
combine the two quote list components but it became awkward due to the
overlapping heading on the non-photo variant, so I've kept them separate.