Prevent cookie acceptance on privacy page with query param
On the privacy pages (cookie, cookie
acceptance, privacy policy) we should never
show the cookie acceptance dialog as it
obscures the content. This usually is the
case, however, when there is a query
parameter in the URL the logi fails to match
as we are matching a path against a path with
a query string.

Match on path without query string.