Ignore anchor tags with no `href`
I'm not sure why this happens but if an `inset_text` component specified in
markdown has a `href` with a jump link, so `<a href="#example">example</a>`
then Nokogiri doesn't parse it correctly and winds up thinking the anchor is an
`id` attribute instead of the `href` attribute (which it thinks is `nil`).

Working around it for now by ignoring anchor links with a `nil` `href`
attribute.