Tweak spacing of action container
This is an awkward one; the way it was previously was fine for short headings
but longer headings ate up vertical padding. The easiest way around it I can
see for now is to have an absolute background behind the heading thats the same
colour as the box below and half height.