Add salaries and benefits test variants
We are running an A/B test for a new salaries and
benefits landing page; to be able to split the
traffic weneed variants of the main salaries and
benefits page for 'search' traffic and 'social'
traffic.

These are temporary and will be removed after the
test has finished, so the content is just duplicated
for ease.