Downgrade Knapsack
We were using a free tier of Knapsack Pro but
they've recently asked us to move onto a paid plan.
Either their policy has changed or we have exhausted
the free tier limits.

Downgrade to regular Knapsack, which still provides
parallelisation, just not quite as optimised as the
pro version.