Support CORS preflight requests
A CORS preflight request is a CORS request that checks to see if the CORS
protocol is understood and a server is aware using specific methods and
headers.

It is an OPTIONS request, using three HTTP request headers:
Access-Control-Request-Method, Access-Control-Request-Headers, and the
Origin header.

We currently respond to all OPTIONS requests with a 404; instead, we can
include the `rack-cors` gem to correctly service these requests.