No-index blog tag pages
The blog tag pages have the same intro paragraph
and could be considered duplicate content
according to Google Search Console. We have also
not reviewed the tags yet so there are likely
improvements that need to be made in this area. In
the meantime we don't want to index these pages
with search engines.