Enable Google Optimize on variant test pages
Currently we only fire Google Optimize on the original page and not the variant
(as the traffic is all sent to the original page and redirected to the variant
from there).

We're seeing oddness around far more traffic going to the original than the
variant (although GA shows more of a split) so we're assuming the Optimize
snippet also needs to execute on the variant page in order for it to register
correctly.