Remove blur from Optimize experiment
Currently, when a user visits a page running an A/B
test if they have not accepted cookies the content
behind the cookie dialog is blurred. This is to
ensure they can't see the original page if they then
get redirected to the variant page on accepting
cookies/opting-in to the experiment.

We think this isn't a great user experience and -
given most A/B tests are the same or similar for
above the fold content - we would rather show the
original briefly and switch out to the variant
without the blurring effect.