Add test to check for cached dynamic pages
When we add a form to a static page it is still automatically cached by the
`rack-page_caching` gem unless we explicitly list it as dynamic in the
`PagesController`. This is open to human error and can result in a form going
live that ends up being cached and therefore the CSRF token will be incorrect
(as the cached version will expire).

Add a test to check all pages that contain POST method forms are listed as
dynamic and excluded from the cache.

Stub the privacy policy requst for all content specs so we no longer need to
exclude certain pages (that otherwise would fail when they try and request the
privacy policy).