Automatically exclude pages from the cache
Currently we maintain a manual array of page paths that we want to exclude from
the static page cache. These tend to be pages that contain a form with a post
request method (if we cache these pages the CSRF token is also cached and
becomes invalid, causing the form submission to fail). Going forward we want to
automate this process as its currently prone to human error.

Add middleware to check if a response contains a form and - if it does - remove
the cached version of the page that is created earlier up the middleware
stack.v