Refactor steps into view component
The steps component is currently just a partial within the steps layout. Going
forward we want to use this component in other locations and also have a
non-numerical version that displays checkmarks in place of the step numbers.

Refactor steps into a view component. Update to support numerical/non-linear
presentation styles. Update template handler to allow it to be included in
other markdown files. Tweak the padding to match the latest design (no left
padding on tablet/mobile viewports).