Configure postgres and redis
Each environment can either create its own services or link to shared
instances

We are in the process of migrating the TTA service into the GiT
website/codebase. As part of this we will be migrating the feedback form from
the TTA service, which is backed by the shared Postgres database.