Fix cache deletion for home page
We auto-delete the page cache if a form is detected. When doing
this we use the request path to find the cached filename. The
homepage is unique because the path is not the same as the cached
page filename (the path is `/` and the filename is `/index.html`).
We need to account for this explicitly to ensure the cached file
is removed when the homepage contains a form.