Support 'course full' on internship providers page
We want to highlight to candidates when a provider course is 'full'. To do
this we are going to accept a new 'full' column on the spreadsheet that must
contain a boolean value. When full, we will remove the contact details and
place a 'Course full' status at the top of the card.