Improve page cache exclusion logic
We had an issue in production whereby the home page was being cached
after we added a form to it, resulting in the CSRF token becoming
invalid.

To avoid this we added middleware to prevent caching of any pages
contaiing a form, but this contained a workaround for the root path,
which is a special case whereby the path does not match the cached
filename (`/` resolves to `/index`).

We added a quick fix for this but it wasn't very robust; instead, this
commit queries the page caching gem to obtain the cached page filename
for the given request path.