Fix Turbolinks page change/scroll issue
On some pages if you scroll down and click a link you will
be taken to the page and the scroll position will be
maintained (instead of scrolling back to the top of the new
page). This bug was introduced when we migrated from
Webpacker to Shakapacker.

The cause appears to be Turbolinks reloading the
`application.js` asset on these page changes. It only
happens when the destination page uses a different layout
containing different Javascript assets; in our case going
from a normal content page to a page containing a form is
the main culprit as we pull in an additional
`govuk_frontend` pack on these pages. Its not clear why
Turbolinks then reloads the `application` pack as well.

We likely want to migrate away from Turbolinks to Turbo in
the future, but we've tried that in the past and it causes a
separate set of headaches with our analytics. A fix for now
is to ensure all pages load the same set of Javascript
packs.