Ensure no direct links to TTA service
As we are migration from the Get an adviser service to a adviser sign up in the
GiT website we need to ensure all TTA service links go to `/tta-service` so
that we can capture and redirect them appropriately.

Remove all absolute links to the Get an adviser service and add a test to
ensure no others are added.