Consolidate TeachingSubject presentation
We currently retrieve teaching subjects from the API in a
multitude of different ways. We want to be consistent in
how we retrieve subjects and also in how we
filter/display them.

Centralise the teaching subject queries to the
`Crm::TeachingSubject` model and update all references to
go through this.