Disable Turbolinks on TTA service paths
Currently, the links to `/tta-service` are handled
by Turbolinks, which results in the page URL being
`/tta-service` after the redirect instead of
`teacher-training-advisers/sign_up/identity`. We
want to make sure the URL updates to the correct
path after redirecting so that A/B tests and Google
Analytics are correct.

Disable Turbolinks on `/tta-service` paths.